Finalised Assassin AI (Important info in desc.)
I've made sure the assassin is fully functional.

See testroom_assassin.bsp for an example of them in action.

See testroom_assassin.vmf to see how to correctly set them up in a level.

Use the enemy disadvantage nodes sparingly otherwise the Assassin may become stuck or confused. You should still try to get good coverage with them however so there'll always be 1 or 2 disadvantage points near the player when fighting the assassin. Don't use them exclusively on places you want them to jump or they'll rarely ever be on the ground which isn't fun. Try to put a few on the ground but still in places where the player will be disadvantaged if the enemy can get there (such as flanks).