Dynamic NPC spawning system is go!
Yup!

Check testmap_dynamicnpc.bsp for a basic example of how it works.

Check testmap_dynamicnpc.vmf for an example of how you can set up the logic and entities!

Check dynamic_npcs for a preview of how this could be applied in a gameplay setting to enhance replayability!

Coming soon: Dynamic point_template! Spawn (or don't!) any entity!