Implemented Dynamic Template Spawner
I've done it again!

The template spawner works almost identically to the NPC spawner, but can be used with ANY ENTITY.

Check testmap_dynamicitems.bsp for a basic example of how it works.

Check testmap_dynamicitems.vmf for an example of how you can set up the logic and entities!

Check dynamic_items for a preview of how this could be applied in a gameplay setting to enhance replayability!