Fix precache error when Cremator fires
This error was being caused by a discrepancy in the original leak code and the updated code. The Cremator originally uses sprites/lgtning.vmt when firing however in the new code he uses the sprite defined by sk_cremator_immolator_beamsprite (which is now sprites/physbeam.vmt). This was causing a problem because the precache code was still only precaching sprites/lgtning.vmt.

To fix this, I made sure that the precache code now precaches the material pointed to by sk_cremator_immolator_beamsprite.