Nerfed combine balls & melee vs Cremator
Before, the Cremator was instantly dying to combine balls (this was also very buggy). It was also possible to quite easily kill him with melee (Cremator has 200 health, crowbar does 10 dmg/hit, 200 / 10 = 20 hits to kill Cremator)

To rectify this, I've added special checks in the Cremator's code to see if he's being damaged by a combine ball or by a weapon with the "club" damage type. I've also modified the combine ball code to make it explode on contact with the Cremator.