First iteration of CGuard patriarch behaviour
This behaviour means that after taken a certain amount of damage, the CGuard will cloak and retreat to heal. This is based on the behaviour of the Patriarch in KF1.

It's current locked behind a ConVar (g_cguard_patriarch) because it doesn't function correctly. The cloaking is too slow, the CGuard doesn't heal and the CGuard doesn't cloak.

This broken first iteration of the behaviour has given me an idea though. Stay tuned.