Implemented hotswap prototype
It's currently just slapped onto the pistol's secondary fire. To enable it use the command pistol_hotswap 1.

Right clicking will hotswap you with whatever NPC you're looking at. Right now that just means a position swap. In the future I'll try to get it working with viewangles as well.

The temporary sound is STOCK AS FUCK. I hope Omni gets a good laugh out of it.