Moved hotswap to concommand, added keybind, now copies NPC angles
Woo! Hotswap is more or less fully functional now.  The option for the keybind can be found under "Gillespie Pass" in the options menu.

It's heavily WIP but the basic functionality is there. Right now you can hotswap with stuff you REALLY shouldn't be able to like gunships, story characters, etc.

As with every new feature I implement make sure to bugtest it for me and make an issue on Github for any problems you find.