Implement Game Instructor
The Game Instructor is the system used in games such as Left 4 Dead and Portal 2 to show onscreen hints. I won't go into detail about the specifics of the more complicated stuff for this system because all you guys need to know is that you can set up GI hints in your map using the new entity, env_instructor_hint.